replace console.log on catch with throw
