+++ remix/app/routes/newsletters.tsx
@@ -1,16 +1,9 @@
 import type { User } from "@prisma/client";
 import type { LoaderArgs } from "@remix-run/node";
-import {
-  Form,
-  isRouteErrorResponse,
-  useLoaderData,
-  useRouteError,
-} from "@remix-run/react";
+import { Outlet, useLoaderData } from "@remix-run/react";
 import { authenticator } from "~/config/auth.server";
 import type { UserProfile } from "~/types";
 import { db } from "~/config/db.server";
-
-import { SocialsProvider } from "remix-auth-socials";
 
 export const loader = async ({ request }: LoaderArgs) => {
   const userRequest = (await authenticator.isAuthenticated(request)) as any;
@@ -27,7 +20,7 @@ export const loader = async ({ request }: LoaderArgs) => {
       where: { email: userProfile.email },
     });
   } catch (err) {
-    console.log(err);
+    throw err;
   }
 
   return { user };
@@ -40,27 +33,7 @@ export default function Index() {
     <div>
       <h1>Newsletter Page</h1>
       <p>{user?.name}</p>
-    </div>
-  );
-}
-
-export function ErrorBoundary() {
-  const error = useRouteError();
-
-  if (isRouteErrorResponse(error) && error.status === 403) {
-    return (
-      <div className="error-container">
-        <p>You must be logged.</p>
-        <Form method="post" action={`/auth/${SocialsProvider.GOOGLE}`}>
-          <button>Login with Google</button>
-        </Form>
-      </div>
-    );
-  }
-
-  return (
-    <div className="error-container">
-      Something unexpected went wrong. Sorry about that.
+      <Outlet />
     </div>
   );
 }
